refactor(linker): isolate symlink mutations
